+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - promote inactivity based session from beta
 - run angular migrations `ngclass-to-class` + `style`
 - a11y: improvements to custom plugins modal (from #2677) (@rfiorentino1)
+- add bridge filter dropdown to accessories page
 
 ### Other Changes